feat: 도메인 객체를 엔티티로 변경
