+++ src/main/java/org/sopt/post/controller/dto/request/UpdatePostRequest.java
@@ -1,15 +1,12 @@
 package org.sopt.post.controller.dto.request;
 
-import java.util.List;
-
 import org.sopt.post.code.FailureCode;
 import org.sopt.post.exception.InvalidContentException;
 import org.sopt.post.exception.InvalidTitleException;
 
 public record UpdatePostRequest(
 		String title,
-		String content,
-		List<String> hashtags
+		String content
 ) {
 	public UpdatePostRequest {
 		this.validate(title, content);
