+++ src/main/java/org/sopt/post/controller/dto/response/PostDetailResponse.java
@@ -1,7 +1,5 @@
 package org.sopt.post.controller.dto.response;
 
-import java.util.List;
-
 public record PostDetailResponse(
 		long id,
 		String title,
@@ -10,7 +8,6 @@ public record PostDetailResponse(
 		String author,
 		String createdAt,
 		boolean isAnonymous,
-		List<String> hashtags,
 		int likeCount,
 		int commentCount,
 		int scrapCount
