+++ src/main/java/org/sopt/post/service/dto/command/CreatePostCommand.java
@@ -1,32 +1,27 @@
 package org.sopt.post.service.dto.command;
 
-import java.util.List;
-
-import org.sopt.post.domain.BoardType;
+import org.sopt.post.entity.BoardType;
 
 public record CreatePostCommand(
 		String title,
 		String content,
 		String author,
 		BoardType boardType,
-		boolean isAnonymous,
-		List<String> hashtags
+		boolean isAnonymous
 ) {
 	public CreatePostCommand(
 			String title,
 			String content,
 			String author,
 			String boardType,
-			boolean isAnonymous,
-			List<String> hashtags
+			boolean isAnonymous
 	) {
 		this(
 				title,
 				content,
 				author,
 				BoardType.from(boardType),
-				isAnonymous,
-				hashtags
+				isAnonymous
 		);
 	}
 }
