+++ src/main/java/org/sopt/post/service/dto/command/UpdatePostCommand.java
@@ -1,10 +1,7 @@
 package org.sopt.post.service.dto.command;
 
-import java.util.List;
-
 public record UpdatePostCommand(
 		String title,
-		String content,
-		List<String> hashtags
+		String content
 ) {
 }
