+++ src/main/java/org/sopt/post/service/dto/information/PostListInfo.java
@@ -2,18 +2,8 @@ package org.sopt.post.service.dto.information;
 
 import java.util.List;
 
-import org.sopt.post.domain.Post;
-
 public record PostListInfo(
 		List<PostSummaryInfo> posts,
 		long totalCount
 ) {
-	public static PostListInfo of(List<Post> posts, int totalCount) {
-		return new PostListInfo(
-				posts.stream()
-						.map(PostSummaryInfo::from)
-						.toList(),
-				totalCount
-		);
-	}
 }
