+++ src/main/java/org/sopt/post/service/dto/information/PostSummaryInfo.java
@@ -2,8 +2,7 @@ package org.sopt.post.service.dto.information;
 
 import java.time.LocalDateTime;
 
-import org.sopt.post.domain.BoardType;
-import org.sopt.post.domain.Post;
+import org.sopt.post.entity.BoardType;
 
 public record PostSummaryInfo(
 		long id,
@@ -15,16 +14,4 @@ public record PostSummaryInfo(
 		String author,
 		BoardType boardType
 ) {
-	public static PostSummaryInfo from(Post post) {
-		return new PostSummaryInfo(
-				post.getId(),
-				post.getTitle(),
-				post.getContent(),
-				post.getLikeCount(),
-				post.getCommentCount(),
-				post.getCreatedAt(),
-				post.getAuthor(),
-				post.getBoardType()
-		);
-	}
 }
